Interactive visualisation of spatial proteomics data

These functions allow one to explore spatial proteomics data interactively.

pRolocVis(object, app = "explore", fcol = "markers", ...)

pRolocVis_aggregate(
  object,
  fcol = "markers",
  groupBy,
  fig.height = "700px",
  nchar = 25,
  ...
)

pRolocVis_compare(
  object,
  fcol = "markers",
  classProfiles = FALSE,
  fig.height = "400px",
  nchar = 25,
  ...
)

pRolocVis_explore(
  object,
  fcol = "markers",
  classProfiles = FALSE,
  fig.height = "700px",
  nchar = 25,
  ...
)

Arguments

object

An instance of class MSnSet, or an MSnSetList of length 2 if using "compare" application.

app

The type of application requested: "explore" (default), "compare" or "aggregate". See description below.

fcol

The feature meta-data label (fData column name) to be used for colouring. Default is "markers". If set to NULL, no annotation is expected. For the "compare" app, a character vector of length 2 is allowed to set different labels for each dataset. If only one label is specified, then this single label will be used to identify the annotation column in both datasets. Please see example herein.

...

Additional parameters passed to plot2D for the "explore" (such as the dimensionality reduction technique, and methods), "compare" apps. For the "aggregate" app this is for additional parameters to be passed to combineFeatures.

groupBy

The feature meta-data label (fData column name) to be used for summarising the features to be combined.

fig.height

Height of the figure.

nchar

Maximum number of characters of the subcellular class names, before their names are truncated. Default is 25.

classProfiles

A logical indicating if a tab displaying individual class profile plots should be displayed. Default is FALSE.

Details

The function pRolocVis is a wrapper for pRolocVis_pca, pRolocVis_compare. and pRolocVis_aggregate. These Shiny apps allow to explore and analyse interactively spatial proteomics data.

The explore Shiny app allows exploration of quantitative data (1) visually through a projection of the dataset, (2) protein profiles, and (3) a searchable feature data table, allowing visualisation of sets of proteins of interest.

The compare Shiny app is meant for comparing protein localisation between two conditions, or two different experiments, replicates etc.

The aggregation Shiny app displays a scatter plot of the maximum or mean distances within each feature (e.g. protein group) according to its components (e.g. peptides) defined by the groupBy argument. A PCA plot of the components is also displayed. It can be used for visualising peptides, PSMs or any other features defined in the feature data of the MSnSet and their distributions.

See also

Author

Lisa Breckels, Thomas Naake and Laurent Gatto

Examples

library("pRoloc")
library("pRolocdata")
#> 
#> This is pRolocdata version 1.35.0.
#> Use 'pRolocdata()' to list available data sets.

## Load the Explore app
data(hyperLOPIT2015)
if (interactive()) {
  pRolocVis(hyperLOPIT2015)
  pRolocVis(hyperLOPIT2015, method = "t-SNE")
  ## store the t-SNE coords and pass a matrix to pRolocVis
  xx <- plot2D(hyperLOPIT2015, method = "t-SNE")
  pRolocVis(xx, method = "none", methargs = list(hyperLOPIT2015))
}

## Load the Compare app
data("hyperLOPITU2OS2018")
data("lopitdcU2OS2018")
xx <- MSnSetList(list(hyperLOPITU2OS2018, lopitdcU2OS2018))
if (interactive()) {
  pRolocVis(xx, app = "compare", fcol = c("markers", "final.assignment"))
}

## Visualise the location and distribution of peptides per protein group
data("hyperLOPIT2015ms2psm")
if (interactive()) {
  ## Combine PSM data to peptides
  hl <- combineFeatures(hyperLOPIT2015ms2psm,
                        groupBy = fData(hyperLOPIT2015ms2psm)$Sequence,
                        method = median)
  ## Visualise peptides according to protein group
  pRolocVis(hl, app = "aggregate", fcol = "markers",
            groupBy = "Protein.Group.Accessions")
}